* [core]: fix drawing with Oyranos colours
- speed up transformation by grouping the CIE*xy background
- set correct pixel layout for oyImage_Create()

+++ icc_draw.h
@@ -86,11 +86,11 @@ public:
   bool zeige_raster;               // Schalter für Darstellung der Rasterlinien
 
   // Wertebereiche
-  double min_x, min_y, max_x, max_y;
+  float min_x, min_y, max_x, max_y;
 
 private:
-  double raster_wert_x_;           // Rasterweite in Benutzereinheiten
-  double raster_wert_y_;
+  float raster_wert_x_;           // Rasterweite in Benutzereinheiten
+  float raster_wert_y_;
   void   zeichneRaster_(void);     // Darstellung des Raster mit Werten
 
   // CIExy Raster
@@ -111,14 +111,14 @@ private:
   int n_speicher;
 
   // einige Koordinaten
-  inline int xNachBild (double val)
-                 { return (int)(((double)xO + (val-min_x)*breite
+  inline int xNachBild (float val)
+                 { return (int)(((float)xO + (val-min_x)*breite
                                               /(max_x-min_x)) + 0.5); }
-  inline int yNachBild (double val)
-                 { return (int)(((double)yO - (val-min_y)*hoehe
+  inline int yNachBild (float val)
+                 { return (int)(((float)yO - (val-min_y)*hoehe
                                               /(max_y-min_y)) + 0.5); }
-  inline double bildNachX (int val) { return ((val-min_x-xO)*(max_x-min_x)/breite); }
-  inline double bildNachY (int val) { return ((yO-val-min_y)*(max_y-min_y)/hoehe); }
+  inline float bildNachX (int val) { return ((val-min_x-xO)*(max_x-min_x)/breite); }
+  inline float bildNachY (int val) { return ((yO-val-min_y)*(max_y-min_y)/hoehe); }
 };
